Computing TZF for Braid Matroids
Expanding the recurrence relation for braid matroids by summing over
all integer partitions. It seems hard to generalize this recurrence
relation to a closed form ... but that is no reason not to compute some.
(Or give up!)

tzf_braid() computes the TZF for B_n, the nth braid matroid, dynamically
by computing the TZF for B_m for all m < n. These are stored in the
list braid_tzf.

robert.miranda@yale.edu